% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.CMA.R
\name{estimate.CMA}
\alias{estimate.CMA}
\title{Estimates correlation and multiplicity adjusted confidence bands}
\usage{
estimate.CMA(boot.grid, est = NULL, se = NULL, ind = NULL, alpha = 0.05)
}
\arguments{
\item{boot.grid}{An N x length(ind) matrix of replicates.}

\item{est}{The original estimate to perform the correction around. If NULL (default) then the row-wise mean of boot.grid is used.}

\item{se}{The original standard error estimate to perform the correction with. If NULL (default) the row-wise sd of boot.grid is used.}

\item{ind}{The grid points which the N functions are observed at. Ff NULL (default) seq(0,1, length = nrow(boot.grid) is used).}

\item{alpha}{specifies the significance level for estimated bands i.e. 1-alpha. Defaults to 0.05/}
}
\value{
A dataframe containing ind, value, se, the CMA correction factor, and upper and lower confidence intervals.
}
\description{
Estimates correlation and multiplicity adjusted confidence bands for a set of bootstrap replicates of a functional linear model coefficient.
}
