% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.boots.R
\name{fit.boots}
\alias{fit.boots}
\title{Refits a functional linear model to bootstrap samples}
\usage{
fit.boots(model, boot.list, ncores = 4)
}
\arguments{
\item{model}{A pffr model object.}

\item{boot.list}{A list of dataframes to re-estimate the model.}

\item{ncores}{Specifies the number of cores to use in parallel processing. Defaults to 4.}
}
\value{
A list of lists. Each list object contains 2 items 'orginal' which is the original coefficient and standard error estimates, and 'boots' which is a matric of bootstrap replicates.
}
\description{
Takes a functional linear model generated by refund::pffr and re-estimates the model coefficients on a set of bootstrap samples.
}
