% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.boot.list.R
\name{level.boot.list}
\alias{level.boot.list}
\title{Generate a list of id level bootstrap samples}
\usage{
level.boot.list(df, id, B)
}
\arguments{
\item{df}{The dataframe to be re sampled.}

\item{id}{The variable in the dataframe to be re-sampled}

\item{B}{The number of bootstrap samples to generate.}
}
\value{
A list of dataframes.
}
\description{
Re-samples data at the id level to form a list of dataframes.May be used to re-sample whole participants or groups.
}
