% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.boot.list.R
\name{simple.boot.list}
\alias{simple.boot.list}
\title{Generate a list of observation level bootstrap samples}
\usage{
simple.boot.list(df, B)
}
\arguments{
\item{df}{The dataframe to be re sampled.}

\item{B}{The number of bootstrap samples to generate.}
}
\value{
A list of dataframes.
}
\description{
Re-samples data at the observation level to form a list of dataframes.
}
